# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/roles/{id}": {
      "patch": {
        "operationId": "MembersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新角色权限（仅超管；超级管理员角色不可改）",
        "tags": [
          "rbac"
        ]
      }
    },
    "/members": {
      "get": {
        "operationId": "MembersController_members",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "成员列表",
        "tags": [
          "rbac"
        ]
      },
      "post": {
        "operationId": "MembersController_createMember",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "邀请制建号（含客户账户）：仅超管，强校验角色↔scope 匹配 + scopeId 存在",
        "tags": [
          "rbac"
        ]
      }
    },
    "/members/{id}": {
      "patch": {
        "operationId": "MembersController_updateMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新成员角色 / 停用（防自我提权）",
        "tags": [
          "rbac"
        ]
      }
    },
...
  },
  "info": {
    "title": "网易有道 CPS 平台 API",
    "description": "账户/鉴权 · RBAC · 审计 · 业务管理 后端契约",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateRoleDto": {
        "type": "object",
        "properties": {}
      },
      "CreateMemberDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}